requirements.txt
pytest ~= 7.4.2
selenium ~= 4.13.0
webdriver-manager ~= 4.0.1
allure-pytest ~= 2.13.2